Add files via upload
Lab 4 involved the creation of a multitude of several programming methods such as the creation of a for loop outputting values 1-100 determining their even/odds properties, the implementation of an if-else statement regarding the selection of an animal determined by a temp. integer and finally the slight correction of code written in certain programs that tested our abilities to not only find problems but resolve them efficiently.